ci: initial set for testing
